requirements.txt
pandas
streamlit
langchain-core
scikit-learn
plotly
langchain
langgraph
langchain-openai
langchain-experimental
python-dotenv
